Split util.h to more concrete util headers
