+++ src/commands/command_parser.h
@@ -27,7 +27,7 @@
 
 #include "parse_util.h"
 #include "status.h"
-#include "util.h"
+#include "string_util.h"
 
 template <typename Iter>
 struct MoveIterator : Iter {
