+++ src/common/io_util.h
@@ -20,23 +20,12 @@
 
 #pragma once
 
-#define __STDC_FORMAT_MACROS
+#include <netinet/in.h>
 
-#include <arpa/inet.h>
-
-#include <cctype>
-#include <chrono>
-#include <cinttypes>
-#include <memory>
-#include <string>
-#include <utility>
-#include <vector>
-
-#include "solarisfixes.h"
 #include "status.h"
 
 namespace Util {
-// sock util
+
 sockaddr_in NewSockaddrInet(const std::string &host, uint32_t port);
 Status SockConnect(const std::string &host, uint32_t port, int *fd);
 Status SockConnect(const std::string &host, uint32_t port, int *fd, uint64_t conn_timeout, uint64_t timeout = 0);
@@ -50,32 +39,9 @@ int GetPeerAddr(int fd, std::string *addr, uint32_t *port);
 int GetLocalPort(int fd);
 bool IsPortInUse(int port);
 
-// string util
-Status DecimalStringToNum(const std::string &str, int64_t *n, int64_t min = INT64_MIN, int64_t max = INT64_MAX);
-Status OctalStringToNum(const std::string &str, int64_t *n, int64_t min = INT64_MIN, int64_t max = INT64_MAX);
-const std::string Float2String(double d);
-std::string ToLower(std::string in);
-bool EqualICase(std::string_view lhs, std::string_view rhs);
-void BytesToHuman(char *buf, size_t size, uint64_t n);
-std::string Trim(std::string in, const std::string &chars);
-std::vector<std::string> Split(const std::string &in, const std::string &delim);
-std::vector<std::string> Split2KV(const std::string &in, const std::string &delim);
-bool HasPrefix(const std::string &str, const std::string &prefix);
-int StringMatch(const std::string &pattern, const std::string &in, int nocase);
-int StringMatchLen(const char *p, int plen, const char *s, int slen, int nocase);
-std::string StringToHex(const std::string &input);
-std::vector<std::string> TokenizeRedisProtocol(const std::string &value);
-
-void ThreadSetName(const char *name);
 int aeWait(int fd, int mask, uint64_t milliseconds);
 
-template <typename Duration = std::chrono::seconds>
-auto GetTimeStamp() {
-  return std::chrono::duration_cast<Duration>(std::chrono::system_clock::now().time_since_epoch()).count();
-}
-inline uint64_t GetTimeStampMS() { return GetTimeStamp<std::chrono::milliseconds>(); }
-inline uint64_t GetTimeStampUS() { return GetTimeStamp<std::chrono::microseconds>(); }
-// file util
 Status Write(int fd, const std::string &data);
 Status Pwrite(int fd, const std::string &data, off_t offset);
+
 }  // namespace Util
