+++ src/common/thread_util.h
@@ -18,35 +18,10 @@
  *
  */
 
-#include "cron.h"
+#pragma once
 
-#include <gtest/gtest.h>
+namespace Util {
 
-#include <memory>
+void ThreadSetName(const char *name);
 
-class CronTest : public testing::Test {
- protected:
-  explicit CronTest() {
-    cron = std::make_unique<Cron>();
-    std::vector<std::string> schedule{"*", "3", "*", "*", "*"};
-    cron->SetScheduleTime(schedule);
-  }
-  ~CronTest() = default;
-
- protected:
-  std::unique_ptr<Cron> cron;
-};
-
-TEST_F(CronTest, IsTimeMatch) {
-  std::time_t t = std::time(0);
-  std::tm *now = std::localtime(&t);
-  now->tm_hour = 3;
-  ASSERT_TRUE(cron->IsTimeMatch(now));
-  now->tm_hour = 4;
-  ASSERT_FALSE(cron->IsTimeMatch(now));
-}
-
-TEST_F(CronTest, ToString) {
-  std::string got = cron->ToString();
-  ASSERT_EQ("* 3 * * *", got);
-}
+}  // namespace Util
