+++ src/types/redis_stream_base.cc
@@ -22,7 +22,7 @@
 
 #include "encoding.h"
 #include "parse_util.h"
-#include "util.h"
+#include "time_util.h"
 
 namespace Redis {
