+++ src/types/redis_string.cc
@@ -25,6 +25,7 @@
 #include <string>
 
 #include "parse_util.h"
+#include "time_util.h"
 
 namespace Redis {
